feat: move use-filters-injection file && refactor category filter creation
