+++ packages/x-components/src/composables/__tests__/use-filters-injection.spec.ts
@@ -5,7 +5,7 @@ import {
   createHierarchicalFilter,
   createSimpleFilter
 } from '../../__stubs__/filters-stubs.factory';
-import { useFiltersInjection } from '../use-filters-injection';
+import { useFiltersInjection } from '../../x-modules/facets/composables/use-filters-injection';
 
 const Provider = defineComponent({
   name: 'Provider',
