+++ packages/x-components/src/composables/index.ts
@@ -7,4 +7,3 @@ export * from './use-store';
 export * from './use-state';
 export * from './use-getter';
 export * from './use-alias-api';
-export * from './use-filters-injection';
