+++ packages/x-components/src/x-modules/facets/components/lists/sliced-filters.vue
@@ -50,7 +50,7 @@
   import { VueCSSClasses } from '../../../../utils';
   import { facetsXModule } from '../../x-module';
   import { useRegisterXModule } from '../../../../composables/use-register-x-module';
-  import { useFiltersInjection } from '../../../../composables/use-filters-injection';
+  import { useFiltersInjection } from '../../composables/use-filters-injection';
 
   /**
    * Component that slices a list of filters and returns them using the default scoped slot,
