+++ packages/x-components/src/x-modules/facets/composables/index.ts
@@ -1,1 +1,2 @@
 export * from './use-facets';
+export * from './use-filters-injection';
